refactor: Break Destinations List screen into separate widgets
