+++ dristi/lib/src/features/destinations/presentation/widgets/destinations_app_bar.dart
@@ -1,0 +1,79 @@
+import 'package:dristi/src/core/base/base_consumer_stateful_widget.dart';
+import 'package:dristi/src/core/constants/app_values.dart';
+import 'package:dristi/src/core/utils/localization_ext.dart';
+import 'package:dristi/src/features/destinations/presentation/riverpod/destination_provider.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:go_router/go_router.dart';
+
+class DestinationsAppBar extends ConsumerStatefulWidget {
+  const DestinationsAppBar({super.key});
+
+  @override
+  ConsumerState createState() => _DestinationsAppBarState();
+}
+
+class _DestinationsAppBarState
+    extends BaseConsumerStatefulWidget<DestinationsAppBar> {
+  final TextEditingController _searchFieldController = TextEditingController();
+  @override
+  Widget build(BuildContext context) {
+    return SliverAppBar(
+      floating: true,
+      flexibleSpace: FlexibleSpaceBar(
+        background: _buildAppBar(),
+      ),
+      automaticallyImplyLeading: false,
+      expandedHeight: AppValues.dimen_70.h,
+    );
+  }
+
+  Widget _buildAppBar() {
+    ref.watch(destinationsSearchField);
+    final searchFieldNotifier = ref.read(destinationsSearchField.notifier);
+
+    return AppBar(
+      leading: IconButton(
+        icon: const Icon(
+          Icons.arrow_back_ios,
+        ),
+        onPressed: () {
+          context.pop();
+        },
+      ),
+      title: TextField(
+        controller: _searchFieldController,
+        onChanged: (value) {
+          searchFieldNotifier.state = value;
+        },
+        onTapOutside: (event) {
+          FocusManager.instance.primaryFocus?.unfocus();
+        },
+        cursorColor: uiColors.primary,
+        style: appTextStyles.secondaryNovaRegular16,
+        decoration: InputDecoration(
+          hintText: context.localization.searchDestination,
+          suffixIcon: _searchFieldController.text.isNotEmpty
+              ? GestureDetector(
+                  onTap: () {
+                    _searchFieldController.clear();
+                    searchFieldNotifier.state = '';
+                  },
+                  child: const Icon(Icons.clear),
+                )
+              : null,
+        ),
+      ),
+      actions: [
+        GestureDetector(
+          onTap: () {},
+          child: Padding(
+            padding: EdgeInsets.symmetric(horizontal: AppValues.dimen_10.w),
+            child: const Icon(Icons.more_vert),
+          ),
+        ),
+      ],
+    );
+  }
+}
